feat : local 회원가입 모달 분기점 완료
